backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
pydantic
python-dotenv
psycopg2-binary

